+++ site-cookbooks/rdr_memcached_wrapper/recipes/default.rb
@@ -1,0 +1,10 @@
+#
+# Cookbook:: rdr_memcached_wrapper
+# Recipe:: default
+#
+# Copyright:: 2021, The Authors, All Rights Reserved.
+
+memcached_instance 'memcached' do
+  extra_cli_options ['-P /var/run/memcached/memcached.pid']
+  user 'memcache'
+end