+++ site-cookbooks/rdr_redisio_wrapper/attributes/default.rb
@@ -1,0 +1,2 @@
+default['redisio']['package_install'] = true
+default['redisio']['version'] = nil