+++ site-cookbooks/rdr_redisio_wrapper/metadata.rb
@@ -1,0 +1,20 @@
+name 'rdr_redisio_wrapper'
+maintainer 'Ben Dixon'
+maintainer_email 'ben@talkingquickly.co.uk'
+license 'MIT License'
+description 'A lightweight wrapper around the community redisio LWRP\'s for installing Redis server'
+version '0.1.0'
+chef_version '>= 15.0'
+depends 'redisio'
+
+# The `issues_url` points to the location where issues for this cookbook are
+# tracked.  A `View Issues` link will be displayed on this cookbook's page when
+# uploaded to a Supermarket.
+#
+# issues_url 'https://github.com/<insert_org_here>/rdr_redisio_wrapper/issues'
+
+# The `source_url` points to the development repository for this cookbook.  A
+# `View Source` link will be displayed on this cookbook's page when uploaded to
+# a Supermarket.
+#
+# source_url 'https://github.com/<insert_org_here>/rdr_redisio_wrapper'