+++ site-cookbooks/rdr_redisio_wrapper/recipes/default.rb
@@ -1,0 +1,14 @@
+#
+# Cookbook:: rdr_redisio_wrapper
+# Recipe:: default
+#
+# Copyright:: 2021, The Authors, All Rights Reserved.
+
+apt_repository 'redis' do
+  uri 'ppa:redislabs/redis'
+end
+
+apt_update 'update'
+
+redisio_install "redis" do
+end