+++ site-cookbooks/rdr_users_wrapper/metadata.rb
@@ -1,0 +1,20 @@
+name 'rdr_users_wrapper'
+maintainer 'Ben Dixon'
+maintainer_email 'ben@talkingquickqly.co.uk'
+license 'MIT License'
+description 'Wrapper cookbook around; https://github.com/sous-chefs/users which will add the users defined in the users data bag'
+version '0.1.0'
+chef_version '>= 15.0'
+depends 'users'
+
+# The `issues_url` points to the location where issues for this cookbook are
+# tracked.  A `View Issues` link will be displayed on this cookbook's page when
+# uploaded to a Supermarket.
+#
+# issues_url 'https://github.com/<insert_org_here>/users_wrapper/issues'
+
+# The `source_url` points to the development repository for this cookbook.  A
+# `View Source` link will be displayed on this cookbook's page when uploaded to
+# a Supermarket.
+#
+# source_url 'https://github.com/<insert_org_here>/users_wrapper'